Add support to get Host_IPA_Limit for AArch64
Host_IPA_Limit is the maximum possible value for
IPA_Bits on the host and is dependent on the CPU capability
and the kernel configuration. The limit can be retrieved using
KVM_CAP_ARM_VM_IPA_SIZE of the KVM_CHECK_EXTENSION ioctl.

This commit adds support for getting Host_IPA_Limit for AArch64.

Signed-off-by: Henry Wang <Henry.Wang@arm.com>